Error handling

+++ BACKEND/src/main/java/org/example/backend/service/BookServiceImpl.java
@@ -1,16 +1,17 @@
 package org.example.backend.service;
 import org.example.backend.dto.response.BookResponseDto;
+import org.example.backend.exception.BookAlreadyExistsException;
 import org.example.backend.mapper.BookMapper;
 import org.example.backend.dto.request.BookRequestDto;
 import org.example.backend.model.Book;
+import org.example.backend.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.example.backend.repository.BookRepository;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.io.IOException;
 import java.util.*;
 import java.util.UUID;
+import org.example.backend.exception.*;
 
 @Service
 public class BookServiceImpl implements BookService{
@@ -33,14 +34,26 @@ public class BookServiceImpl implements BookService{
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+        Optional<Book>existingBook=bookRepository.findByIsbn(requestDto.getIsbn());
+        if(!existingBook.isPresent()){
         Book book = bookMapper.toEntity(requestDto,imageUrl);
         bookRepository.save(book);
-        return bookMapper.toDto(book);
+        return bookMapper.toDto(book);}
+        else{
+            throw new BookAlreadyExistsException("Book already exist");
+        }
     }
 
     @Override
-    public void deleteBook(UUID id){
-        bookRepository.deleteById(id);
+    public String deleteBook(UUID id){
+        Optional<Book>existingBook=bookRepository.findById(id);
+        if(existingBook.isPresent()){
+            bookRepository.deleteById(id);
+        }
+        else{
+            throw new BookNotFoundException("Book does not exists.");
+        }
+        return "Book deleted successfully";
     }
 
     @Override
@@ -48,6 +61,4 @@ public class BookServiceImpl implements BookService{
         List<Book>books=bookRepository.findAll();
         return bookMapper.toDtoList(books);
     }
-
 }
-
